code for 4-d-2

+++ blog/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -10,23 +11,38 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const body = request.body
-  const users = await User.find({})
-  const firstUser = users[0]
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({
+  //     error: 'token missing or invalid'
+  //   })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+  const {title, author, url, likes} = request.body
   const blogObj = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-    user: firstUser._id,
+    title,
+    author,
+    url,
+    likes,
+    user: user._id,
   })
   const savedBlog = await blogObj.save()
-  firstUser.blogs = firstUser.blogs.concat(savedBlog._id)
-  await firstUser.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const blog = await Blog.findById(request.params.id);
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = request.user
+  if (user._id.toString() !== blog.user.toString()) {
+  // if (decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({
+      error: 'unauthorized'
+    })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
